Extract lsystem_bridge.rs, rename fractal_renderer.rs to line_renderer.rs
Move geometry_to_vertices and color_params_from_config out of the wgpu
module into a new lsystem_bridge.rs so line_renderer.rs has no lsystem-core
dependency. Rename FractalPipelineResources → LinePipeline, FractalRenderer
→ GpuContext, and tighten all pub items to pub(crate) throughout.
