+++ CLAUDE.md
@@ -50,9 +50,10 @@ Data flow: `Config` → `ExpandIter` (lazy string rewriting) → `Turtle2D` → 
 |------|------|
 | `main.rs` | Thin native entry that calls `lib.rs::run_native()` |
 | `lib.rs` | Module declarations; `run_native()` builds an `EventLoop<UserEvent>` and calls `run_app`; `#[wasm_bindgen(start)] start()` does the same on web via `EventLoopExtWebSys::spawn_app` |
-| `fractal_renderer.rs` | Toolkit-independent wgpu module. `Transform` — scale + offset GPU uniform type. `FractalRenderer` — owns the wgpu surface; `begin_frame` acquires the next surface texture and `end_frame` submits + presents. `FractalPipelineResources` (pipeline, bind group, vertex buffer, transform uniform, color-params uniform) — `upload()` re-uploads vertices and `ColorParams`; `write_transform()` writes the camera transform every frame; `draw()` issues the line-list draw. On wasm `FractalRenderer` is built asynchronously and delivered via `UserEvent::GpuReady` |
+| `line_renderer.rs` | Toolkit-independent wgpu module. `Transform` — scale + offset GPU uniform type. `GpuContext` — owns the wgpu surface; `begin_frame` acquires the next surface texture and `end_frame` submits + presents. `LinePipeline` (pipeline, bind group, vertex buffer, transform uniform, color-params uniform) — `upload()` re-uploads vertices and `ColorParams`; `write_transform()` writes the camera transform every frame; `draw()` issues the line-list draw. On wasm `GpuContext` is built asynchronously and delivered via `UserEvent::GpuReady` |
+| `lsystem_bridge.rs` | L-system→GPU adapters. `geometry_to_vertices()` converts `Geometry` to a flat `Vertex` array with bounding box (`VertexData`). `color_params_from_config()` maps `LineColorConfig` to the `ColorParams` GPU uniform. |
 | `renderer.rs` | `App` (`ApplicationHandler<UserEvent>`) — owns `Camera`, geometry buffer, side-panel state. Routes `WindowEvent::RedrawRequested` straight to its own renderer; routes everything else through `egui-winit` |
-| `ui.rs` | `UiState` (preset/config state, egui layout including the central fractal canvas via `ui.allocate_painter()`, pan/zoom from the painter `Response`) + `EguiRenderer` (egui context, egui-wgpu integration, single render pass that does both the surface clear and the fractal+egui draw) + `FractalCallback` (per-frame data struct: vertices, transform, needs_upload, color_params) + `impl egui_wgpu::CallbackTrait for FractalCallback` (thin egui adapter that delegates to `FractalPipelineResources::upload/write_transform/draw`) |
+| `ui.rs` | `UiState` (preset/config state, egui layout including the central fractal canvas via `ui.allocate_painter()`, pan/zoom from the painter `Response`) + `EguiRenderer` (egui context, egui-wgpu integration, single render pass that does both the surface clear and the fractal+egui draw) + `FractalCallback` (per-frame data struct: vertices, transform, needs_upload, color_params) + `impl egui_wgpu::CallbackTrait for FractalCallback` (thin egui adapter that delegates to `LinePipeline::upload/write_transform/draw`) |
 | `camera.rs` | `Camera` (pan/zoom state), `compute_transform` |
 | `shader.wgsl` | Vertex shader applies a `Transform` uniform (scale + offset) and computes per-segment color from a `ColorParams` uniform using `vertex_index / 2`; supports solid, gradient, and HSV hue-cycle modes; fragment shader passes the interpolated color through; topology is `LineList` |
