+++ crates/lsystem-app/src/line_renderer.rs
@@ -1,7 +1,6 @@
 use std::sync::Arc;
 
 use bytemuck::{Pod, Zeroable};
-use lsystem_core::{Geometry, LineColorConfig};
 use wgpu::util::DeviceExt;
 use winit::window::Window;
 
@@ -15,49 +14,7 @@ pub(crate) struct Transform {
 #[repr(C)]
 #[derive(Copy, Clone, Pod, Zeroable)]
 pub(crate) struct Vertex {
-    position: [f32; 2],
-}
-
-pub(crate) struct VertexData {
-    pub vertices: Vec<Vertex>,
-    pub bounds_min: [f32; 2],
-    pub bounds_max: [f32; 2],
-}
-
-pub(crate) fn geometry_to_vertices(geometry: &Geometry) -> VertexData {
-    let Geometry::D2 { segments } = geometry;
-
-    let mut min_x = f32::INFINITY;
-    let mut min_y = f32::INFINITY;
-    let mut max_x = f32::NEG_INFINITY;
-    let mut max_y = f32::NEG_INFINITY;
-    let mut vertices = Vec::with_capacity(segments.len() * 2);
-
-    for [a, b] in segments {
-        min_x = min_x.min(a.x).min(b.x);
-        min_y = min_y.min(a.y).min(b.y);
-        max_x = max_x.max(a.x).max(b.x);
-        max_y = max_y.max(a.y).max(b.y);
-        vertices.push(Vertex {
-            position: [a.x, a.y],
-        });
-        vertices.push(Vertex {
-            position: [b.x, b.y],
-        });
-    }
-
-    if min_x.is_infinite() {
-        min_x = -1.0;
-        max_x = 1.0;
-        min_y = -1.0;
-        max_y = 1.0;
-    }
-
-    VertexData {
-        vertices,
-        bounds_min: [min_x, min_y],
-        bounds_max: [max_x, max_y],
-    }
+    pub(crate) position: [f32; 2],
 }
 
 /// Maximum number of line segments that fit in a 256 MiB vertex buffer (wgpu's guaranteed limit).
@@ -67,70 +24,22 @@ pub(crate) const MAX_SEGMENTS: u64 = 268_435_456 / (2 * std::mem::size_of::<Vert
 /// Per-frame color parameters written to the GPU as a uniform.
 /// Layout mirrors `ColorParams` in `shader.wgsl`; padding keeps vec4 alignment.
 #[repr(C)]
-#[derive(Copy, Clone, Pod, Zeroable)]
+#[derive(Copy, Clone, Default, Pod, Zeroable)]
 pub(crate) struct ColorParams {
     /// 0 = solid, 1 = gradient, 2 = hue_cycle
-    mode: u32,
-    total_segments: u32,
-    _pad: [u32; 2],
-    color_start: [f32; 4],
-    color_end: [f32; 4],
-    hue_start: f32,
-    saturation: f32,
-    value: f32,
-    _pad2: f32,
+    pub(crate) mode: u32,
+    pub(crate) total_segments: u32,
+    pub(crate) _pad: [u32; 2],
+    pub(crate) color_start: [f32; 4],
+    pub(crate) color_end: [f32; 4],
+    pub(crate) hue_start: f32,
+    pub(crate) saturation: f32,
+    pub(crate) value: f32,
+    pub(crate) _pad2: f32,
 }
 
-impl Default for ColorParams {
-    fn default() -> Self {
-        color_params_from_config(&LineColorConfig::default(), 0)
-    }
-}
-
-pub(crate) fn color_params_from_config(line: &LineColorConfig, total_segments: u32) -> ColorParams {
-    match *line {
-        LineColorConfig::Solid(c) => ColorParams {
-            mode: 0,
-            total_segments,
-            _pad: [0; 2],
-            color_start: [c[0], c[1], c[2], 1.0],
-            color_end: [0.0; 4],
-            hue_start: 0.0,
-            saturation: 0.0,
-            value: 0.0,
-            _pad2: 0.0,
-        },
-        LineColorConfig::Gradient { start, end } => ColorParams {
-            mode: 1,
-            total_segments,
-            _pad: [0; 2],
-            color_start: [start[0], start[1], start[2], 1.0],
-            color_end: [end[0], end[1], end[2], 1.0],
-            hue_start: 0.0,
-            saturation: 0.0,
-            value: 0.0,
-            _pad2: 0.0,
-        },
-        LineColorConfig::HueCycle {
-            start_hue,
-            saturation,
-            value,
-        } => ColorParams {
-            mode: 2,
-            total_segments,
-            _pad: [0; 2],
-            color_start: [0.0; 4],
-            color_end: [0.0; 4],
-            hue_start: start_hue,
-            saturation,
-            value,
-            _pad2: 0.0,
-        },
-    }
-}
-
-/// GPU resources for fractal rendering.
-pub(crate) struct FractalPipelineResources {
+/// GPU pipeline for rendering colored line-list geometry.
+pub(crate) struct LinePipeline {
     pipeline: wgpu::RenderPipeline,
     uniform_buffer: wgpu::Buffer,
     color_params_buffer: wgpu::Buffer,
@@ -139,7 +48,7 @@ pub(crate) struct FractalPipelineResources {
     vertex_count: u32,
 }
 
-impl FractalPipelineResources {
+impl LinePipeline {
     pub(crate) fn new(device: &wgpu::Device, format: wgpu::TextureFormat) -> Self {
         let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
             label: None,
@@ -299,14 +208,14 @@ pub(crate) enum FrameOutcome {
     Skip,
 }
 
-pub struct FractalRenderer {
+pub(crate) struct GpuContext {
     surface: wgpu::Surface<'static>,
-    pub device: Arc<wgpu::Device>,
-    pub queue: Arc<wgpu::Queue>,
+    pub(crate) device: Arc<wgpu::Device>,
+    pub(crate) queue: Arc<wgpu::Queue>,
     surface_config: wgpu::SurfaceConfiguration,
 }
 
-impl FractalRenderer {
+impl GpuContext {
     pub(crate) async fn new(window: Arc<Window>) -> Result<Self, ()> {
         let size = window.inner_size();
 
@@ -356,15 +265,15 @@ impl FractalRenderer {
         })
     }
 
-    pub fn surface_format(&self) -> wgpu::TextureFormat {
+    pub(crate) fn surface_format(&self) -> wgpu::TextureFormat {
         self.surface_config.format
     }
 
-    pub fn size(&self) -> (u32, u32) {
+    pub(crate) fn size(&self) -> (u32, u32) {
         (self.surface_config.width, self.surface_config.height)
     }
 
-    pub fn resize(&mut self, new_size: winit::dpi::PhysicalSize<u32>) {
+    pub(crate) fn resize(&mut self, new_size: winit::dpi::PhysicalSize<u32>) {
         if new_size.width == 0 || new_size.height == 0 {
             return;
         }
@@ -403,71 +312,5 @@ impl FractalRenderer {
         if reconfigure_after {
             self.surface.configure(&self.device, &self.surface_config);
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use lsystem_core::{Config, generate};
-
-    const EPS: f32 = 1e-5;
-
-    fn close(a: f32, b: f32) -> bool {
-        (a - b).abs() < EPS
-    }
-
-    fn cfg(toml: &str) -> Config {
-        Config::parse(toml).unwrap()
-    }
-
-    #[test]
-    fn empty_geometry_uses_fallback_bounds() {
-        // axiom has no F, so no segments are drawn
-        let geom = generate(&cfg(
-            "name=\"t\"\naxiom=\"A\"\niterations=0\nangle=90.0\nstep=1.0",
-        ));
-        let VertexData {
-            vertices,
-            bounds_min,
-            bounds_max,
-        } = geometry_to_vertices(&geom);
-        assert!(vertices.is_empty());
-        assert!(close(bounds_min[0], -1.0) && close(bounds_min[1], -1.0));
-        assert!(close(bounds_max[0], 1.0) && close(bounds_max[1], 1.0));
-    }
-
-    #[test]
-    fn single_segment_produces_two_vertices_and_tight_bounds() {
-        // "F" at 0 iterations: one segment from (0,0) to (1,0)
-        let geom = generate(&cfg(
-            "name=\"t\"\naxiom=\"F\"\niterations=0\nangle=90.0\nstep=1.0",
-        ));
-        let VertexData {
-            vertices,
-            bounds_min,
-            bounds_max,
-        } = geometry_to_vertices(&geom);
-        assert_eq!(vertices.len(), 2);
-        assert!(close(vertices[0].position[0], 0.0) && close(vertices[0].position[1], 0.0));
-        assert!(close(vertices[1].position[0], 1.0) && close(vertices[1].position[1], 0.0));
-        assert!(close(bounds_min[0], 0.0) && close(bounds_min[1], 0.0));
-        assert!(close(bounds_max[0], 1.0) && close(bounds_max[1], 0.0));
-    }
-
-    #[test]
-    fn bounds_are_tight_over_all_segments() {
-        // "F+F-F": three segments covering x=[0,2], y=[0,1]
-        let geom = generate(&cfg(
-            "name=\"t\"\naxiom=\"F+F-F\"\niterations=0\nangle=90.0\nstep=1.0",
-        ));
-        let VertexData {
-            vertices,
-            bounds_min,
-            bounds_max,
-        } = geometry_to_vertices(&geom);
-        assert_eq!(vertices.len(), 6);
-        assert!(close(bounds_min[0], 0.0) && close(bounds_min[1], 0.0));
-        assert!(close(bounds_max[0], 2.0) && close(bounds_max[1], 1.0));
     }
 }
